+++ packages/e2e-vite/utils/types.ts
@@ -13,20 +13,22 @@ export interface E2EPlaywrightContext {
 }
 
 export interface Evaluator {
-  getTextContent: (selector: string, page?: Page) => Promise<string | Nil>
-  getColor: (selector: string, page?: Page) => Promise<string | Nil>
-  getDisplayStyle: (selector: string, page?: Page) => Promise<string | Nil>
-  getAssetUrl: (selector: string, page?: Page) => Promise<string | Nil>
-  getJustifyContent: (selector: string, page?: Page) => Promise<string | Nil>
+  getTextContent: (selector: string, page?: Page) => Promise<EvaluateResult>
+  getColor: (selector: string, page?: Page) => Promise<EvaluateResult>
+  getDisplayStyle: (selector: string, page?: Page) => Promise<EvaluateResult>
+  getAssetUrl: (selector: string, page?: Page) => Promise<EvaluateResult>
+  getJustifyContent: (selector: string, page?: Page) => Promise<EvaluateResult>
   getElementCount: (selector: string, page?: Page) => Promise<number>
-
-  // Actions
   clearLocalStorage: (page?: Page) => Promise<void>
   inputText: (selector: string, text: string, page?: Page) => Promise<void>
   click: (selector: string, page?: Page) => Promise<void>
+  isImageLoaded: (selector: string, page?: Page) => Promise<EvaluateResult>
 }
+
+export type EvaluateResult = string | boolean | Nil
 
 export type EvaluateType
   = | 'style'
     | 'attribute'
     | 'textContent'
+    | 'isImageLoaded'