+++ packages/eslint-plugin/src/utils.ts
@@ -5,10 +5,10 @@ import type { Rule } from 'eslint'
 import type { VineESLintDocs } from './types'
 import { NS } from '@vue-vine/eslint-parser'
 import { expect } from 'vitest'
-import HTML_ELEMENTS from './data/html-elements.json'
-import MATH_ELEMENTS from './data/math-elements.json'
-import SVG_ELEMENTS from './data/svg-elements.json'
-import VOID_ELEMENTS from './data/void-elements.json'
+import HTML_ELEMENTS from './data/html-elements'
+import MATH_ELEMENTS from './data/math-elements'
+import SVG_ELEMENTS from './data/svg-elements'
+import VOID_ELEMENTS from './data/void-elements'
 
 export interface RuleModule<
   T extends readonly unknown[],
@@ -100,7 +100,7 @@ export const createEslintRule = RuleCreator(
   TMessageIds extends string,
   TDocs extends VineESLintDocs = VineESLintDocs,
 >(
-  { name, meta, ...rule }: Readonly<RuleWithMetaAndName<TOptions, TMessageIds, TDocs>>
+  { name, meta, ...rule }: Readonly<RuleWithMetaAndName<TOptions, TMessageIds, TDocs>>,
 ) => RuleModule<TOptions>
 
 export function notVineCompFn(node: any): boolean {