+++ packages/nuxt-module/test/fixtures/basic/nuxt.config.ts
@@ -1,6 +1,7 @@
 import vueVineNuxtModule from '../../../src/module'
 
 export default defineNuxtConfig({
+  ssr: true,
   modules: [
     vueVineNuxtModule,
   ],