+++ packages/rsbuild-plugin/tsdown.config.ts
@@ -1,6 +1,7 @@
+import type { UserConfig } from 'tsdown'
 import { defineConfig } from 'tsdown'
 
-const tsdownConfig: ReturnType<typeof defineConfig> = defineConfig({
+const tsdownConfig: UserConfig = defineConfig({
   entry: {
     index: './src/index.ts',
   },