+++ packages/tsc/tsdown.config.ts
@@ -1,7 +1,11 @@
+import type { UserConfig } from 'tsdown'
 import { defineConfig } from 'tsdown'
 
-const tsdownConfig: ReturnType<typeof defineConfig> = defineConfig({
-  // Because @volar/typescript used dynamic require
-  format: 'cjs',
+const tsdownConfig: UserConfig = defineConfig({
+  dts: false,
+  outputOptions: {
+    // Because @volar/typescript used dynamic require
+    format: 'cjs',
+  },
 })
 export default tsdownConfig