+++ packages/vscode-ext/CHANGELOG.md
@@ -1,5 +1,11 @@
 # vue-vine-extension
 
+## 1.7.27
+
+### Patch Changes
+
+- Fix language service for broken on concating props type in virtual code.
+
 ## 1.7.26
 
 ### Patch Changes